Refactor graphql schema creation
This is preparation for creating an API for sites and plugins to add
arbitrary types to the schema.
