+++ lib/bootstrap/index.js
@@ -126,7 +126,7 @@ module.exports = async (program, cb) => {
   }
 
   // Create Schema.
-  const schema = await require(`../utils/schema`)()
+  const schema = await require(`../schema`)()
   const graphqlRunner = (query, context) => graphql(schema, query, context, context, context)
 
   // TODO create new folder structure for files to reflect major systems.
