+++ lib/schema/node.js
@@ -1,0 +1,14 @@
+import {
+  nodeDefinitions,
+} from 'graphql-relay'
+
+//const { nodeInterface, nodeField } = nodeDefinitions(
+module.exports = nodeDefinitions(
+  (globalId) => {
+    const { type, id } = fromGlobalId(globalId)
+    return null
+  },
+  (obj) => {
+    return obj.ships ? factionType : shipType
+  }
+)
