+++ package.json
@@ -16,7 +16,7 @@
     "url": "https://github.com/gatsbyjs/gatsby/issues"
   },
   "dependencies": {
-    "async": "^2.0.1",
+    "async": "^2.1.2",
     "babel-core": "^6.14.0",
     "babel-loader": "^6.2.5",
     "babel-plugin-add-module-exports": "^0.2.1",
@@ -28,6 +28,7 @@
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
     "babel-runtime": "^6.11.6",
+    "bluebird": "^3.4.6",
     "boom": "^2.7.2",
     "bundle-loader": "^0.5.4",
     "chalk": "^1.1.3",
@@ -85,7 +86,7 @@
     "react-router-scroll": "^0.3.2",
     "remark": "^6.0.1",
     "remark-html": "^5.0.1",
-    "remark-parse": "^2.0.1",
+    "remark-parse": "^2.0.2",
     "sanitize-html": "^1.13.0",
     "slash": "^1.0.0",
     "static-site-generator-webpack-plugin": "^3.0.0",
@@ -112,11 +113,9 @@
     "babel-plugin-lodash": "^3.2.9",
     "babel-plugin-transform-flow-strip-types": "^6.14.0",
     "babel-register": "^6.14.0",
-    "bluebird": "^3.4.6",
     "cheerio": "^0.22.0",
-    "eslint": "^3.5.0",
+    "eslint": "^3.8.1",
     "eslint-config-airbnb": "^12.0.0",
-    "eslint-plugin-ava": "^3.0.0",
     "eslint-plugin-flow-vars": "^0.5.0",
     "eslint-plugin-flowtype": "^2.18.2",
     "eslint-plugin-import": "^2.0.1",
